Fix 2130 current setting and CS current scaling calculation

+++ src/modules/utils/motordrivercontrol/drivers/TMC21X/TMC21X.h
@@ -271,7 +271,7 @@ public:
      * \param current the maximum motor current in mA
      * \sa getCurrent(), getCurrentCurrent()
      */
-    void set_current(unsigned int current);
+    void set_current(uint16_t current);
 
     /*!
      * \brief set standstill motor current
